Move UI sketch controls to separate module
